requirements.txt
aiohttp==3.7.4.post0
cchardet==2.1.7
aiodns==3.0.0
aiohttp-cors==0.7.0
aiohttp-session==2.9.0
aioredis==1.3.1
asyncpg==0.23.0
aiohttp-sqlalchemy==0.15.3
uvloop==0.15.2
python-dotenv==0.17.1